route_engine/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
prometheus-client==0.20.0
redis==5.0.7
psycopg2-binary==2.9.9
networkx==3.3
numpy==1.26.4
shapely==2.0.4
